+++ src/reuse/__init__.py
@@ -19,14 +19,19 @@ import gettext
 import logging
 import os
 import re
+from dataclasses import dataclass, field
 from typing import NamedTuple, Set
 
+try:
+    from importlib.metadata import PackageNotFoundError, version
+except ImportError:
+    from importlib_metadata import PackageNotFoundError, version
+
 from boolean.boolean import Expression
-from pkg_resources import DistributionNotFound, get_distribution
 
 try:
-    __version__ = get_distribution(__name__).version
-except DistributionNotFound:
+    __version__ = version("reuse")
+except PackageNotFoundError:
     # package is not installed
     __version__ = "1.1.2"
 
@@ -82,13 +87,17 @@ _IGNORE_SPDX_PATTERNS = [
 # Combine SPDX patterns into file patterns to ease default ignore usage
 _IGNORE_FILE_PATTERNS.extend(_IGNORE_SPDX_PATTERNS)
 
-#: Simple structure for holding SPDX information.
-#:
-#: The two iterables MUST be sets.
-SpdxInfo = NamedTuple(
-    "SpdxInfo",
-    [("spdx_expressions", Set[Expression]), ("copyright_lines", Set[str])],
-)
+
+@dataclass(frozen=True)
+class SpdxInfo:
+    """Simple class holding SPDX information"""
+
+    spdx_expressions: Set[Expression] = field(default_factory=set)
+    copyright_lines: Set[str] = field(default_factory=set)
+    contributor_lines: Set[str] = field(default_factory=set)
+
+    def __bool__(self):
+        return any(self.__dict__.values())
 
 
 class ReuseException(Exception):